AI transports (#6)
* Create new BTree for transports

* Get units to board haulers/transports

* Rename BTree transport -> hauler

Make the name consistent with the rest of the code

* Move haulers to pick up hauled units

* Simplify BTrees

Remove redundant condition leaves
Set city/unit separately from setting move

* Refactor BTree

Break setting move to city into separate components

* Separate BTree move towards/onto hauler

* Move units to coast for hauling

* Pause exploration when there are units to haul

* Transports move to pick up and drop off units

* Move around region with hauler to unhaul

* Fix unit plan path corner case

* Automate hauling units to enemy cities

* Break region automation into separate files

Maybe this helps with debugging

* Only unhaul to enemy cities

* Require valid units

Sometimes they disappear in battles

* Attacking/destroying haulers destroys all hauled

* Use local variable for BT exploration state

Reduces use of blackboard and chance of colliding global variables

* Organize BT code into folders

Keeps hauler/hauled code in dedicated directories for easier management

* Unhaul units to unexplored regions

* Look in neighbor coordinates to unhaul units

* Fix BT exploration loop

Empty the destination path before exploring

* only unhaul units if the hauler has no path moves

* Disable diagonal paths

They are making units get stuck

* Fix bug starting non-water maps

There are no approaches or approach segments

* Fix AI unit unhaul to attack enemy

* Add unit block state

This allows us to track blocked moves. The first use case is for
transports trying to move to an occupied square. They can not attack, so
they are blocked.

* Update to Godot 4.3

Ignore new warnings about unused signals

* Fix debug errors after godot/beehave updates

* Adapt movement test to new algorithm

* Fix ai movement test description

* Protect against AI overthinking

If it plans a move but the move fails, it gets stuck in a loop. If this
happens, stop thinking and stop moving, so we can continue with the
game.

* Fix hauled unit targeting hauled unit

AI can see tanks on transports, but can't get to them. So don't target
them for un-hauling.

* Exclude non-city-capture units from tally

We determine whether a team has lost based on whether it can continue
fighting. If it can no longer take cities, it has lost.

* Verify tanks can attack transports in cities

* If only AI and neutral cities remain, AI wins

In the GUT for this, instead of dumb sleeps, test/watch for signals

* Standardize GUT scenario names

Avoid using the player-oriented scenarios

* Set units from code on ocean test map

Then we can re-use the map for various tests

* updated GUT to latest

* Fix transport path finding

It was finding a path through neutral or hostile units
Since it can not attack, nor take cities, this was breaking the AI

* Add names to all cities

* Create an obstacles manager

* Fix obstacle test referents to points

This uses a grid of points, not node2d positions

* Fix grid occupied logic

Add path length test
Check point objects for occupancy

* Change "my_team" to "team_name"

* Detect obstacles in path

Starting with hauled units finding path to hauler

* Detect obstacles during path finding

* Fix obstacle and turn race conditions

If unit is still being removed, that team is not yet done
If obstacle is being deleted, do not consider it an obstacle

* Fix explore bug on small island

Tank would not unhaul onto a small island in order to explore. Thus
transport was waiting forever for tank to unhaul.

* Fix region approach edge case

Region approaches must exclude tiles that are only adjacent to port
cities

This is because tanks can not move from a transport directly into an
un-owned city

* Remove diagonal approaches for unhauling

* Fix premature end of game for human team

If human team has at least one unit capable of capturing a city, game
should continue.

* Add scenario with enemy transport

* Convert mp3 sound effects to wav

Trying to fix "click" audio effects on html5

* Normalize volumes, and fade before stop

This is an attempt to fix webexport audio glitching

* Deconflict scenario name prefix

* Fix game over miscalculation

If there are no offensive units on either side, and one side still has
cities, that side is the winner

* Fix test scenarios in lobby

If we're building/releasing, exclude test scenarios

* Switch starting cities to closed

They switch to "closed" only when conquered, so we should start them in
that state.

---------

Co-authored-by: Kurt <gtd@m.greenfly.io>